+++ src/libstd/unstable/mod.rs
@@ -22,7 +22,6 @@ pub mod simd;
 pub mod lang;
 pub mod sync;
 pub mod mutex;
-pub mod atomics;
 pub mod raw;
 
 /**
